+++ sdk/servicebus/azure-servicebus/azure/servicebus/aio/_servicebus_sender_async.py
@@ -4,7 +4,7 @@
 # --------------------------------------------------------------------------------------------
 import logging
 import asyncio
-from typing import Any, TYPE_CHECKING, Union, List
+from typing import Any, TYPE_CHECKING, Union, List, Iterable
 
 import uamqp
 from uamqp import SendClientAsync, types
@@ -139,15 +139,15 @@ class ServiceBusSender(BaseHandlerAsync, SenderMixin):
         except Exception as e:
             raise MessageSendFailed(e)
 
-    async def schedule(self, message, schedule_time_utc):
-        # type: (Union[Message, BatchMessage], datetime.datetime) -> List[int]
-        """Send Message or BatchMessage to be enqueued at a specific time.
+    async def schedule(self, messages, schedule_time_utc):
+        # type: (Union[Message, Iterable[Message]], datetime.datetime) -> List[int]
+        """Send Message or multiple Messages to be enqueued at a specific time.
         Returns a list of the sequence numbers of the enqueued messages.
-        :param message: The messages to schedule.
-        :type message: ~azure.servicebus.Message or ~azure.servicebus.BatchMessage
+        :param messages: The message or list of messages to schedule.
+        :type messages: ~azure.servicebus.Message or Iterator[~azure.servicebus.Message]
         :param schedule_time_utc: The utc date and time to enqueue the messages.
         :type schedule_time_utc: ~datetime.datetime
-        :rtype: List[int]
+        :rtype: list[int]
 
         .. admonition:: Example:
 
@@ -160,10 +160,10 @@ class ServiceBusSender(BaseHandlerAsync, SenderMixin):
         """
         # pylint: disable=protected-access
         await self._open()
-        if isinstance(message, BatchMessage):
-            request_body = self._build_schedule_request(schedule_time_utc, *message._messages)
+        if isinstance(messages, Message):
+            request_body = self._build_schedule_request(schedule_time_utc, messages)
         else:
-            request_body = self._build_schedule_request(schedule_time_utc, message)
+            request_body = self._build_schedule_request(schedule_time_utc, *messages)
         return await self._mgmt_request_response_with_retry(
             REQUEST_RESPONSE_SCHEDULE_MESSAGE_OPERATION,
             request_body,
@@ -171,12 +171,12 @@ class ServiceBusSender(BaseHandlerAsync, SenderMixin):
         )
 
     async def cancel_scheduled_messages(self, sequence_numbers):
-        # type: (Union[int, List[int]]) -> None
+        # type: (Union[int, Iterator[int]]) -> None
         """
         Cancel one or more messages that have previously been scheduled and are still pending.
 
         :param sequence_numbers: he sequence numbers of the scheduled messages.
-        :type sequence_numbers: int or list[int]
+        :type sequence_numbers: int or Iterator[int]
         :rtype: None
 
         .. admonition:: Example: