+++ sdk/servicebus/azure-servicebus/tests/test_queues.py
@@ -1088,7 +1088,6 @@ class ServiceBusQueueTests(AzureMgmtTestCase):
                     raise Exception("Failed to receive schdeduled message.")
             
 
-    @pytest.mark.skip("Pending message scheduling functionality")
     @pytest.mark.liveTest
     @pytest.mark.live_test_only
     @CachedResourceGroupPreparer(name_prefix='servicebustest')
@@ -1110,11 +1109,11 @@ class ServiceBusQueueTests(AzureMgmtTestCase):
                     message_id_b = uuid.uuid4()
                     message_b = Message(content)
                     message_b.properties.message_id = message_id_b
-                    tokens = sender.schedule(enqueue_time, message_a, message_b)
+                    tokens = sender.schedule([message_a, message_b], enqueue_time)
                     assert len(tokens) == 2
     
-                messages = receiver.fetch_next(timeout=120)
-                messages.extend(receiver.fetch_next(timeout=5))
+                messages = receiver.receive(max_wait_time=120)
+                messages.extend(receiver.receive(max_wait_time=5))
                 if messages:
                     try:
                         data = str(messages[0])
@@ -1130,7 +1129,6 @@ class ServiceBusQueueTests(AzureMgmtTestCase):
                     raise Exception("Failed to receive schdeduled message.")
             
 
-    @pytest.mark.skip(reason="Pending message scheduling functionality")
     @pytest.mark.liveTest
     @pytest.mark.live_test_only
     @CachedResourceGroupPreparer(name_prefix='servicebustest')
@@ -1146,10 +1144,10 @@ class ServiceBusQueueTests(AzureMgmtTestCase):
                 with sb_client.get_queue_sender(servicebus_queue.name) as sender:
                     message_a = Message("Test scheduled message")
                     message_b = Message("Test scheduled message")
-                    tokens = sender.schedule(enqueue_time, message_a, message_b)
+                    tokens = sender.schedule([message_a, message_b], enqueue_time)
                     assert len(tokens) == 2
     
-                    sender.cancel_scheduled_messages(*tokens)
+                    sender.cancel_scheduled_messages(tokens)
     
                 messages = receiver.receive(max_wait_time=120)
                 try: